+++ include/cfx/vector.h
@@ -11,7 +11,7 @@ extern "C" {
 #define CFX_VEC_MIN_CAP 16
 
 typedef struct {
-    uint32_t* data;
+    uint64_t* data;
     size_t size;
     size_t cap;
 } cfx_vec_t;
@@ -19,8 +19,8 @@ typedef struct {
 void cfx_vec_init(cfx_vec_t *v);
 
 int  cfx_vec_reserve(cfx_vec_t *v, size_t need);
-int  cfx_vec_resize(cfx_vec_t *v, size_t sz, uint32_t fill);
-int  cfx_vec_push(cfx_vec_t *v, uint32_t value);
+int  cfx_vec_resize(cfx_vec_t *v, size_t sz, uint64_t fill);
+int  cfx_vec_push(cfx_vec_t *v, uint64_t value);
 
 void cfx_vec_clear(cfx_vec_t *v);
 void cfx_vec_free(cfx_vec_t *v);
