+++ src/vector.c
@@ -11,7 +11,7 @@ void cfx_vec_init(cfx_vec_t *v) {
 int cfx_vec_reserve(cfx_vec_t* v, size_t need) {
     if (need < v->cap) return 0;
     
-    const size_t max_size = SIZE_MAX / sizeof(uint32_t);
+    const size_t max_size = SIZE_MAX / sizeof(uint64_t);
     if (need > max_size) return -1;
 
     size_t new_cap = v->cap ? 2 * v->cap : CFX_VEC_MIN_CAP;
@@ -20,16 +20,16 @@ int cfx_vec_reserve(cfx_vec_t* v, size_t need) {
         new_cap *= 2;
     }
 
-    void *p = realloc(v->data, new_cap * sizeof(uint32_t));
+    void *p = realloc(v->data, new_cap * sizeof(uint64_t));
     if (!p) return -1;
 
-    v->data = (uint32_t*)p;
+    v->data = (uint64_t*)p;
     v->cap = new_cap;
     printf("reserved cap %zu\n", new_cap);
     return 0;
 }
 
-int cfx_vec_resize(cfx_vec_t* v, size_t sz, uint32_t fill) {
+int cfx_vec_resize(cfx_vec_t* v, size_t sz, uint64_t fill) {
     if (sz > v->cap) {
         if (cfx_vec_reserve(v, sz) != 0) return -1;
     }
@@ -42,7 +42,7 @@ int cfx_vec_resize(cfx_vec_t* v, size_t sz, uint32_t fill) {
     return 0;
 }
 
-int cfx_vec_push(cfx_vec_t* v, uint32_t x) {
+int cfx_vec_push(cfx_vec_t* v, uint64_t x) {
     if (v->size == v->cap) {
         if (cfx_vec_reserve(v, v->size + 1) != 0) return -1;
     }
@@ -65,7 +65,7 @@ void cfx_vec_clear(cfx_vec_t* v) {
 void cfx_vec_print(const cfx_vec_t* v) {
     printf("v: %p, size: %zu\n", v, v->size);
     for (size_t i = 0; i < v->size; ++i) {
-        printf("%u ", v->data[i]);
+        printf("%llu ", v->data[i]);
     }
     printf("\n");
 }
